+++ intelmq/tests/bots/parsers/cymru/test_cap_program_new.py
@@ -203,11 +203,21 @@ EVENTS = [{'time.source': '2019-03-22T11:18:52+00:00',
            'time.source': '2020-07-09T03:40:15+00:00',
            'source.account': 'pm',
            },
+          {'classification.type': 'scanner',
+           'classification.identifier': 'darknet',
+           'time.source': '2020-10-08T02:21:26+00:00',
+           'protocol.transport': 'gre',
+           },
+          {'classification.type': 'scanner',
+           'classification.identifier': 'darknet',
+           'time.source': '2020-10-15T09:22:10+00:00',
+           'protocol.transport': 'ipv6-nonxt',
+           },
           ]
 
 # The number of events a single line in the raw data produces
-NUM_EVENTS = [1, 1, 1, 1, 1, 2, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
-              1, 1, 10, 1, 1, 1, 1, 1]
+NUM_EVENTS = (1, 1, 1, 1, 1, 2, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
+              1, 1, 10, 1, 1, 1, 1, 1, 1, 1)
 RAWS = []
 for i, line in enumerate(RAW_LINES[3:]):
     for count in range(NUM_EVENTS[i]):